+++ src/myplayhrt.c
@@ -679,10 +679,10 @@ int main(int argc, char *argv[])
                      fprintf(stderr, "playhrt: Cannot stat shared memory %s.\n", fnames[i-optind]);
                      exit(24);
                  }
-                 size = sb.st_size - sizeof(int);
+                 size = sb.st_size - sizeof(long);
              }
              /* map the memory (will be on page boundary, so 0 mod 8) */
-             mems[i-optind] = mmap(NULL, sizeof(int)+size,
+             mems[i-optind] = mmap(NULL, sizeof(long)+size,
                              PROT_WRITE | PROT_READ, MAP_SHARED, fd[i-optind], 0);
              if (mems[i-optind] == MAP_FAILED) {
                  fprintf(stderr, "playhrt: Cannot map shared memory.");
@@ -732,7 +732,7 @@ int main(int argc, char *argv[])
           /* get lock */
           sem_wait(*sem);
           /* find length of relevant memory chunk */
-          flen = *((int*)(*mem));
+          flen = (int)(*((long*)(*mem)));
           if (flen == 0) {
               /* done, unlink semaphores and shared memory */
               fname = fnames;
@@ -747,7 +747,7 @@ int main(int argc, char *argv[])
               exit(0);
           }
           /* write shared memory content hardware buffer  */
-          ptr = *mem + sizeof(int);
+          ptr = *mem + sizeof(long);
           sz = 0;
 
           /* loop until shared memory file read */
